DRF-JWT-Djoser/manager/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
